Miscellaneous compatibility changes
One change forces us to step to a newer version of epicsdbbuilder.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/support/epics_device@166365 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ src/epics_extra.c
@@ -1,6 +1,7 @@
 /* Extra support built on top of epics_device. */
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdarg.h>
@@ -297,9 +298,10 @@ void database_add_macro(const char *macro_name, const char *format, ...)
     va_start(args, format);
 
     /* Arbitrary limit for the length of a single macro definition. */
-    const int macro_size = 512;
+    const size_t macro_size = 512;
     char macro_string[macro_size];
-    ASSERT_OK(vsnprintf(macro_string, macro_size, format, args) < macro_size);
+    ASSERT_OK(
+        vsnprintf(macro_string, macro_size, format, args) < (int) macro_size);
     va_end(args);
 
     if (database_macros)
